requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.3.6
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Deprecated==1.2.15
Django==4.2.11
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.5
django-simple-history==3.4.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.5.0
future==1.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
line-bot-sdk==3.14.2
multidict==6.1.0
mysqlclient==2.2.6
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==5.9.6
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.0
yarl==1.18.3
